Improve Compatibility of .OPTIONS OUTPUT and FFT
This improves the compatibility of .OPTIONS OUTPUT INITIAL_INTERVAL
and .OPTIONS OUTPUT OUTPUTTIMEPOINTS with both .MEASURE and
.OPTIONS FFT FFT_ACCURATE.

.OPTIONS OUTPUT OUTPUTTIMEPOINTS was made to work with FFT_ACCURATE=1
by modifying Transient::tranopOutputs() to call the function
outputManagerAdapter_.tranOutput() with skipPrintLineOutput = true,
for the case where 0 is NOT in the list in OUTPUTTIMEPOINTS.  This
allows the Measure objects to be updated at t=0.  More importantly,
it also allows the FFTAnalysis objects to set the required
sample-time breakpoints, since that is done at t=0.

.OPTIONS OUTPUT INITIAL_INTERVAL can not be made compatible with
FFT_ACCURATE=1, without a lot of changes to the tran outputter.
(The Bugzilla entry explains why.)  So, instead the function
FFTMgr::fixupFFTParameters() modified to reset FFT_ACCURATE=0, if
.OPTIONS OUTPUT INITIAL_INTERVAL is used, and also emit a warning
message.  This works fine, and again the Bugzilla entry explains
why it works.

SON Bugs 1280, 1327 and 1331
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1331