Update CMake Windows and AMD install instructions
I took the AMD build instructions out of the Windows-specific area and
put them into a subsection of the TPL build instructions. I also
modified the CmakeLists.txt file for building the AMD library, itself.
It is now virtually identical to the process used in the superbuild (the
two could probably be merged some time in the future).

Finally, I made a small tweak to the Cygwin build instructions.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1