Change default behavior of random operators.
Issue #178.
"Change default behavior of random operators (AGAUSS, GAUSS, etc) to
return the mean when not doing sampling."

This commit removes all the code that supports the old behavior.  The
old behavior pre-dates both the various UQ methods in Xyce, as well as
the new expression library.

Now, if there are no sampling commands in the netlist, random operators
just return the mean.  To reproduce the old behavior, one can add a
.SAMPLING command to the netlist but specify a single sample point.

This commit also removes the command line  option:

Xyce -hspice-ext random

This will break around 4-5 regression tests that were designed to
support the old default behavior, as well as at least one that exercises
the command line option.