Add boolean funcs such as "getVariableDependent"
Issue #35
"Reduce netlist setup time for modern PDKs"

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

This commit is part of an work to see if I can make the use of
global parameters more efficient.  I've trying to reduce the number of
tree traversals performed by the expression library.    One place I
noticed (while runing profiler "Instruments") where traversals happen is
whenever the CircuitContext or DeviceEntity calls "getVariables" or
"getVoltageNodes" on an expression, the expression library will often
then perform a traversal to gather this information.  It only does this
if the expression was updated after the last traversal, so it doesn't
happen as often is it might have.

But, if the only thing that the function is attempting to do is determine if
a parameter expression depends on (say) voltage nodes, as part of an error
trap, this is a silly thing to do here.  That error trap doesn't really
need to know exactly which voltage nodes are present in the expressions.
Arguably it might need them to emit an informative error message, but it
can make that call inside of an if-statement.

The possible drawback of this idea is that some expressions are going to
have to do such a traversal eventually.  So, there is that.  I may just
be pushing these traversals to a later stage of the setup.   These
traversals are only necessary for expressions that have external
dependencies, and they only have to be done once. It remains to be seen
if this will save time or not.

In this commit, I am only *adding* the accessor functions to the
expression library that allows one to query (with a boolean)
if an expression depends on a various entities.   In a later commit
I'll add code that actually uses them in the CircuitContext and
DeviceEntity.