Add usage of expression dependency booleans.
Issue #35
"Reduce netlist setup time for modern PDKs"

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

In my previous commit I added new functions to the expression library
that allows it to return booleans indicating external dependences, such
as for global_params, voltage nodes, device currents, and the like.

This commit adds a bunch of code to actually use these functions during
the circuit parsing and setup phases.  This modified code is in the
Circuit Context class and the device Entity class.

At the moment, this code is kind of messy, with large blocks commented
in and out.  I will clean this up later.

I have debugged this code so that all of the serial nightly test suite
passes.  This took a lot of work to track everything down.  The one
exception to this is the BUG_1222_SON/translated_vco/us_vco.cir test
case.  That tests case still times out. (i.e. takes way too long).  It
takes way to long because on this branch, all .params are treated as
global_params, and apparently even with my recent changes on this
branch, this is too expensive.