Add a basic sanity check for Trilinos
If Trilinos was not properly configured in certain ways, the test for
LOCA in NOX would fail. This can lead to confusion as to the real cause
of the failure. This commit adds a simple test to see if a file that
includes `Teuchos_SerialDenseMatrix.hpp` can be compiled against the
given Trilinos installation. (That header file should always be in a
Trilinos build that can support Xyce.) This way, the configure will fail
before the LOCA test and give an appropriate error message.

While working through this, I found the `HAVE_LIBPARMETIS` variable had
not been added to the CMake build system. It should now do the "right
thing" if Trilinos is built with awareness of ParMETIS.

Finally, I cleaned up some some comments and a long error message line.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1