Add basic Trilinos sanity check to autotools build
In commit 6eed2f299, @jcverle added a "basic sanity check" to the
CMake build to check that Trilinos has been built in a way that can
work with Xyce.  This basic sanity check is merely to check that
Teuchos_SerialDenseMatrix.hpp is present in the Trilinos install and
can be compiled.

I am adding the identical check to the autotools build.

The idea here is to make the configure process fail early when given an
incompatible build of Trilinos, before it hits the probe for LOCA
headers.  Teuchos_SerialDenseMatrix.hpp should always be present in a
build of Trilinos that is compatible with Xyce.

Not in fact in any way a part of
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1, but an attempt to
make the two build systems do the same kinds of tests as long as we
have both.