Reorder Trilinos content checks in autotools
Our autotools build has been checking carefully for the presense of
EpetraExt_AMD_CrsGraph.h if the user has requested AMD support (or
rather, not disabled AMD support).  After that it checks whether
Teuchos has been compiled with COMPLEX support.

In a recent commit I just added a further sanity check that
Teuchos_SerialDenseMatrix.hpp exists.

There is a certain illogic to the order in which these tests are
performed.  The HAVE_TEUCHOS_COMPLEX is pretty fundamental and should
go first, the Teuchos_SerialDenseMatrix.hpp is similarly fundamental
and indicates an incompatible build of Trilinos that could never be
used with Xyce.  The AMD one is not a show stopper --- the user can
use the Trilinos build just fine if they rerun configure with AMD
disabled.

So the tests that indicate a completely unusable Trilinos install
should come first, and the one that could be worked around only happen
if the other two indicate that it's usable.