Add Metric/Measure Definitions for .FFT and .MEASURE FFT
This commits provides the detailed definitions of the ENOB, SFDR,
SNDR, SNR and THD metrics/measures for .FFT and .MEASURE FFT lines.
It covers the cases of FREQ, FMAX and/or FMIN being used on .FFT lines.
It also covers the cases of MAXFREQ, MINFREQ and/or NBHARM being
used on .MEASURE lines.

SON Bugs 1280, 1327 and 1330
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1330