Make top-level context param act like global_param
Issue #167.
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

I spent a lot more time on this than expected before realizing that I
could address this issue with a small amount of new code.

Note, .param and .global_param were very similar before this commit, in
that they could both be changed during a simulation.  The only
difference was that the device package was only tracking globals in its
map.  So, any analysis that called DeviceMgr::setParam could only modify
globals.  This included DC/STEP/SAMPLING/EMBEDDEDSAMPLING/PCE and LOCA.

To address this, I originally experimented with elevating ALL top-level .param
to .global_param status.  This "worked" but caused a few problems.  One was
that there are many circumstances under which globals are implicitly not
allowed.  That was fixable.   A more problematic issue was that processing
of globals is simply slower, and I struggled to find a quick way to fix this.
When something is a .param, it has the opportunity to be declard as "constant"
during the setup phase, and this prevents any more operations to be performed
on it.  No more checks, or re-evalautions, or diagnostics.  Apparently, for
really large circuits with modern PDKs this efficiency really matters.

I made several completely separate attempts to fix the above problems,
but eventually realized that the best thing to do was to keep everything
as it is, but then elevate .params to .global_param status only when
absolutely necessary.  So, if a param is invoked by .STEP, or any other
parameter-modifying analysis, then it is elevated to a global, but
otherwise it is left alone.  Doing this only required that I add a
single function in the context, called CircuitContext::categorizeParams.
This function is called from the CircuitBlock::parseNetlistFilePass1
function, after the handleLinePass1 function call (which parses all the
parameters and "dot" lines), but before the
circuitContext_.resolve function call, which resolves parameters in
different ways depending on if they are global nor not.

This commit accomplishes nearly everything I hoped to for this issue, so
I plan to declare it fixed and implement tests for it.