Add support for exclamation point in param names
Issue #191.
"Support the exclamation point character in parameter names, in the
expression library"

This involves the "ID" macro at the top of the expression lexer file.
I had previously left the "!" character out of this macro out of concern for
conflicts with math operators.  But it needs to be added to support "Y"
device parameters.

I initially added the exclamation point directly to the ID macro, but I
suspected this would cause a conflict with the "not equal" operator.
My suspicion proved correct.  So, the final correct fix in this commit
adds a second macro, ID2, which includes the exclamation point,
but makes it so that '!' can only be in a name if it is NOT
the beginning or ending character of the name.