Add unit tests for '!' character in param names
Issue #191.
"Support the exclamation point character in parameter names, in the
expression library"

This commit adds a single new test to both programs, which tests the
if a parameter name containing '!' can be properly parsed.

This commit also renames a test to make more sense.

Finally, it also fixes a mistake in the parserUnitTest program, which
was causing a bunch of tests to be incorrectly skipped.  I made a
mistake in applying preprocessor macros at some point, and inadvertantly
commented out a bunch of tests.