Small Update to .FFT and .MEASURE FFT Text in Guides
This commit adds N() as a valid output variable to the .FFT
section of the Users' Guide.  This is for consistency with the
existing text for .FOUR.  It also notes that multi-terminal
lead currents (e.g., IC()) are not supported for the FIND measure
for FFT measure mode, in the Reference Guide.

SON Bugs 1280, 1327 and 1332
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1332