Update for AC .sens objective funcs w/expressions
Issue #81.
"Use the new expression library for AC sensitivity objective functions"

I had thought this bug was done but I overlooked a lot of the
documentation.  I had done the user guide but not the release notes or
the reference guide.

This commit adds text about this issue to the reference guide, so it is
now up-to-date.  This commit also adds AC objective functions to release
notes, in the "interface improvements" section.