Extend .param usage to .sens and .data
Issue #167.
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also:
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

I was working on documenting this issue and then realized that I
neglected to support .SENS and .DATA, both of which can act to tweak
.global_param parameter values.  So, while most parts of Xyce can use
.param exactly as a .global_param, these two capabilities were not able
to.

Fortunately, fixing this was pretty easy, and just required a few edits
to the new "categorizeParams" function in the circuit context class.
This commit adds those edits and fixes this oversight.

I have created a couple of new tests to test this out.