Update .global_param/.param guidance
Issue #167.
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also:
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

In a handful of places, it was claimed that .param couldn't be changed
via commands like .step.  This commit fixes it.  I think I've taken care
of all the places that had incorrect information about .param and
.global_param.