Fix/improve Trilinos CMake initial cache files
The "trilinos-config" files in the cmake/trilinos directory were not
consistent with each other in terms of the features they enabled. I
figured out a way to have the MPI-oriented file include the regular
build file so the synchronization issue will not occur again.

This commit also makes some small changes to comments and instructions
in various other files. Of note is that I added a reference to the new
Xyce package in Spack.