Add error trap for GMIN stepping.
Issue #174.  Kind of.
"Document all the PCE related methods in the guides for release 7.3"

I've been working on documenting PCE methods, including this one.  There
are multiple PCE-based methods in Xyce, and this one is
the least "ready for prime time", but I wanted to get a better
sense of where it was.

One thing I discovered (or rather, remembered) is that the fully
intrusive PCE method will fail with a seg fault if Xyce
attempts to do GMIN stepping.  In retrospect
this is not surprising as there were similar issues earlier with
embedded sampling.  Basically any numerical method that requires
tweaking the linear system requires support functions in the relevant
linear algebra classes, and the PCE related classes lack some of those
functions.

There isn't any chance of all of that being fixed prior to the 7.3
release, and given that embedded sampling w/ NISP basically provides the
same capability much more robustly, fixing this isn't a super-high
priority at the moment.  But I wanted to make the seg fault go away,
especially since Xyce always attempts GMIN stepping if the Newton solver
fails the DCOP.

The main thing this commit does is to implement the function
PCEBuilder::vnodeGIDVec.  But the implementation doesn't do anything
other than return a fatal error, which is better than a seg fault.

I've also added comments in the code indicating what functions are
missing, and will need to be implemented eventually.