Make ADMS device letter upper case
An open source user of Xyce/ADMS has pointed out that Xyce/ADMS
currently generates a "device name" that is lower cased rather than
upper cased like all the native devices.  This name only appears in
the device count summary, but the difference in case stands out as an
oddity.  So for example, in one of our VBIC13 tests, the summary looks
like this:

***** Device Count Summary ...
       I level 1 (Independent Current Source) 1
       q level 11 (VBIC 1.3 3T)               1
       R level 1 (Resistor)                   1
       V level 1 (Independent Voltage Source) 1
       ----------------------------------------
       Total Devices                          4

instead of this:
***** Device Count Summary ...
       I level 1 (Independent Current Source) 1
       Q level 11 (VBIC 1.3 3T)               1
       R level 1 (Resistor)                   1
       V level 1 (Independent Voltage Source) 1
       ----------------------------------------
       Total Devices                          4

This commit fixes the template that generates the name to uppercase
the SPICE device letter before creating the name that will appear
here.