Fix so .param expressions don't need curly braces
Issue #195.
"Consider making expressions that don't have any spaces in them (such as
agauss by itself) not need curly braces or quotes."

This has come up various times with modern netlists and PDKs.
In Xyce expressions have historically always required that
they be delimeted using curly braces. In more recent years single
quotes have also been allowed as a delimeter for expressions.

There are many use cases for other simulators in which expressions
are specified without any delimeters.   Some of these have been
automatically handled in Xyce for a long time, as part of the resolution
of bug 1692 on bugzilla years ago.  But that issue only addressed
implicit expressions specified on device instance lines.  It
intentionly excluded any "dot" lines, so it has never applied to .param
or .global_param.

It occured to me recently that we could probably support  (more broadly)
implicit expressions without delimeters in Xyce, with just a little
work in the parser.  As it turns out, I was correct about that.

This commit implements fix for this issue, for expressions that are
assigned to .param and .global_param parameters.   This fix is entirely
in the N_IO_ParsingHelpers.C file, and modifies the
extractParamData and extractGlobalParamData functions.  They now call
a new helper function, combineParamValueFields.  This new helper
function figures out which tokens should be combined (if any) to create
the full expression string for a parameter.  It then also adds curly
braces to the resulting combined string.

Also included in this commit is a modification to an error message in
the device block, to make it more informative.