Make "m/M" for meters be a valid unit in expr lexer
This came up while I worked on Issue #195:
"Consider making expressions that don't have any spaces in them
(such as agauss by itself) not need curly braces or quotes."

As part of my work for that, more parameters get parsed by the
expression library than before.  And in one test case, the
expression library could not handle this combination of scalar
suffix and units: "um", for micro meters or microns.

So, I added "mM" to the macro for UNITS in the lexer and this
made the problem go away.  I then set about creating unit tests
and broke a few things.  I discovered then (I had forgotten)
that "mM" is already in the SUFFIX macro, meaning that there
is a potential conflict.  The suffix version of "m" means
"scale by 1.0e-3", while of course the units version of "m"
means "don't do anything as this is just a unit".

The expression library already supported the scale suffix
version of "m".   But I've decided to keep the new UNITS as-is,
because the way the lexer is written the suffix will always take
precendent over the unit, and I think that is the most reasonable
choice anyway.  The expression library has supported "m" as a
scalar since the beginning; I had just neglected to test it in my
unit tests.

This commit adds a bunch of new unit tests for "m", both as a
suffix and as a unit.  When "m" is by itself, it is a scalar
(mili-).  When it is combined with another scalar, as in the
"um" example, then it is a unit (meters).