Add issue 157,167 and bugzilla 1190 to fixed table
This commit adds the following to the fixed defects table, in a single
entry:

Bugzilla 1190:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

Issue #157
"Make the use of specials (TEMP, TIME, VT, FREQ) in .param less
restrictive."

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

The bugzilla issue serves as an umbrella for the to gitlab issues.
Issue #157 has been closed already, but it belongs with the other two,
so I'm including it in this commit.