Remove defect about non-constant .param
This was caught by Rich.  In the known defects section, I had not
removed this entry, pertaining to bug 1037 on bugzilla.

"1037-SON: The use of non-constant values in .PARAM
statements may lead to unexpected results"

Since Issue #157 and Issue #167 have been fixed, this is no longer a
valid warning.  Removing from the table.