Get rid of incorrect warning message
This is related to Issue #81.
"Use the new expression library for AC sensitivity objective functions"

Setting up AC objective functions with the new expression library
required that I create a new parameter called ACOBJFUNC.

There are a bunch of packages in Xyce which receive the netlist options
for sensitivities.  All of them potentially emit errors if they contain
unrecognized parameters.   I had updated all but one of them to
understand that ACOBJFUNC was a valid parameter.  The one that I did NOT
update was the function in the Fourier manager.  That is fixed now.