Add minor fix and error trap to PCE methods
This came up as I was documenting the PCE methods in the user guide.

Issue #174
"Document all the PCE related methods in the guides for release 7.3"

As I was creating an example circuit that combined embedded sampling with
non-intrusive spectral projection(NISP), I noticed that it was emitting
a random seed to the terminal output.

This is normally done with embedded sampling when using embedded
sampling to do LHS or MC sampling, which depend on random numbers.  But
it should not be necessary when doing a quadrature-based PCE method,
which is what NISP is.  If doing .EMBEDDEDSAMPLING and the parameter
projection_pce=true is set on the .options EMBEDDEDSAMPLES line, then
the quadrature-based method is being used.

Quadrature based PCE methods the function evaluations to be
performed by Xyce at quadrature points, rather than at randomly sampled
points.  (regression PCE can use randomly sampled points).  The
quadrature points are set by the order of the scheme and the number of
uncertain parameters, etc, so they are completely deterministic.

In the code it looked as though I had intended to have an if-statement
that prevented the seed from being set up when using quadrature, but I
got the logic wrong.  And, of course, I copied that logic so it appeared
in both the N_ANP_Sampling and N_ANP_EmbeddedSampling classes.  It is
fixed now.

I also discovered that if I accidentally specified zero parmaeters in
the netlist, and I was using the new expression based random operators
to set up the problem, that there was no failsafe to exit with an error.
And, making it worse, at least one of the quadrature PCE setup functions
hangs if there are zero input parameters.  This commit adds a failsafe
error check right before the Stokhos objects are set up, and exits with
a fatal error if the number of uncertain parameters is zero.

This commit also removes one block of commented-out "if 0"
code that should not have still be there.