Add more PCE documentation
Issue #174
"Document all the PCE related methods in the guides for release 7.3"

This commit mostly adds a Gilbert cell mixer example, and compares
several PCE methods.  It is copied out of the final SAND report from
that project, but with the text substantially rewritten to be consistent
with a User guide.

Still a long way to go.

One possible concern, I added a schematic of the gilbert cell which
relies upon the tikz package.  Hopefully that doesn't cause problems.