Fix Error Handling for Issue #81 (ACOBJFUNC)
This commit fixes the error handling of the ACOBJFUNC and OBJVARS
parameters on .SENS lines for .AC analyses.  It covers the case
where neither is given on the .SENS line.  Moving all of that error
checking to AC::doInit() prevents a segfault later in the function
AC::solveSensitivity_().

This is related to:
 https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/81
"Use the new expression library for AC sensitivity objective functions"