Fix users guide errors
The new pce section included use of a \varPsi character, which is only
defined if one includes the amsmath package.  amsmath was included by
very old versions of the SAND report package, but not by the current
one.  I have added a usepackage command to make the thing work.

I have also fixed a couple of minor typos that I spotted when verifying
that adding amsmath didn't break things.