Fix for using random operators in .param
Issue #200.
"Using random operator like AGAUSS in a .param does not result in the
parameter being elevated to .global_param"

Also related to Issue #174
"Document all the PCE related methods in the guides for release 7.3"

For several of the PCE examples I've attempted to put together in the user
guide, this issue came up.  Basically, this is one use case that slipped
thru the cracks when I worked on Issue #167.

This commit adds another block of code to the
CircuitContext::categorizeParams function.  That function checks if any
UQ commands are in the netlist, and if there are, it checks the
unresolvedParam object to see if any of them are dependent on a random
operator such as aunif or agauss.  If they are, then this is another
criteria for elevating them from .param to .global_param status.