Add new error trap, and output basis size
Related to Issue #174
"Document all the PCE related methods in the guides for release 7.3"

Also kind of related to Issue #200.
"Using random operator like AGAUSS in a .param does not result in the
parameter being elevated to .global_param"

I previously added an error trap for quadrature PCE, which caused Xyce
to exit with a fatal error if the vector of uncertain input parameters
was empty.  It needed to do this prior to setting up the various Stokhos
objects, because one of them would go into an infinite loop if the size
was zero.

I discovered today that there was a similar problem for regression PCE.
But in the case of regression PCE, having an empty list of uncertain
parmaeters caused a seg fault.  So, this commit adds another error trap,
that will exit with a fatal error using the same error message.

The above was mostly happening when I attempted to do UQ analysis when
the uncertain parmaeters were set via random operators in .param
parameter expressions.  That issue is fixed now, but one could still
have an empty set of uncertain params, so these error traps were
important.

This commit also adds code to both sapling classes to output (to
terminal output) the size of the PCE basis.  This is mostly
important for regression PCE, since for that algorithm the user has to
set the number of sample points.  In the way I've set up the regression
PCE, it can only handled an overspecified problem and has to exit with
an error if underspecified.  So the number of sample points must be at
least as big as the basis size.   Also, it is best practice for
regression methods like this to use a number of samples that is at least
2x the basis size.  So if the basis size is 45, then  best practice (for
accuracy, etc) is to set numparams=90.  It is much easier to do this if
Xyce tells you the basis size.

Anyway, the basis size output is now included in the standard terminal
output for embedded sampling & sampling (when using either form of
non-intrusive PCE) and intrusive PCE.