More documenting PCE and cleanup SAMPLING sections
Related to Issue #174
"Document all the PCE related methods in the guides for release 7.3"

This commit incorporates many new examples and a lot of new text.

Also, as I worked on it, I discovered small mistakes in some of the
warning messages in the PCE class.  The warning messages all referred to
sampling, but they should have instead referred to PCE.