Add sections for PCE, and update sampling sections
Issue #174.
"Document all the PCE related methods in the guides for release 7.3"

This updates the reference guide for Issue #174.

Sampling and embedded sampling had been documented in this guide
before now, but the documentation excluded anything to do with
augmentation via non-intrusive PCE.  That was b/c release builds
did not enable stokhos before this release.

Fully intrusive PCE was not documented at all before now, for the same
reasons.  The fully intrusive form of PCE cannot exist without Stokhos
enabled, so it was left out of the guides altogether.