Add mention of support for PCE methods
This  adds mention of all types of PCE methods to the Enhanced
Solver Stability, Performance and Features section.  I put it into 2
bullets, one about non-intrusive methods which augment the 2 kinds of
sampling, and the other about the fully intrusive form of PCE that uses
the .PCE command.

Issue #174.
"Document all the PCE related methods in the guides for release 7.3"