Note that OpenMPI package in latest Ubuntu is OK
Ubuntu had a messed up OpenMPI package for nearly three years, and
included that messed up package in the 18.04 LTS release without
fixing it in subsequent package upgrades.  But Ubuntu 20.04 LTS
finally brought in the fixes that had been in Debian's "testing"
branch for some time, and is no longer broken.

Our dire warnings about Ubuntu being broken in releases "17.10 and
later" are now no longer true, and I've modified the language.

Ubuntu 18.04 LTS is still broken, and is still a supported system
upstream (for at least another few years).  It is also the system that
Sandia is considering as an additional COE, so it is still important
to keep this warning around.