Make Separate Output Files for Continuous Measures
This commit re-factors the output paths for .MEASURE lines to provide
the (default) option for the results for each continuous mode measure
(AC_CONT, DC_CONT, NOISE_CONT and TRAN_CONT modes) to be sent to its
own output file.  The file will be <netlistName>_<measurename>.mt0
for TRAN_CONT measures where measurename will be lower-cased.

For testing purposes (mainly), .OPTIONS MEASURE USE_CONT_FILES=<val>
was added as an option.  The default is 1 (true) for compatibility
with HSPICE.  Setting that option to 0 (false) will send all of
the .MEASURE output to the <netlistName>.mt0 file.  This reduces
the number of output files that need to be verified during the
functional testing of the continuous-mode measure classes.

SON Bug 1274
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274