Some documentation updates for naked expressions
Issue #195
"Make (more) expressions that don't have any spaces in them (such as
agauss by itself) not need curly braces or quotes"

This commit adds information about naked expressions to the release
notes (in the interface improvements section), the user guide, and the
reference guide.

The work on the reference guide section
led me to improving the regression tests for this feature, which in turn
led me to realize my implementation for this issue had overlooked
comparison operators.    That is fixed now.