Fix 195 fix to account for comparison operators
Issue #195
"Make (more) expressions that don't have any spaces in them (such as
agauss by itself) not need curly braces or quotes"

My previous fix was mostly fine, but I overlooked one use case.  That
use case was comparison operators such as ">=".

My previous fix relied on the fact that you can use the equals sign as a
delimeter on .param and .global_param lines.  So far so good, but when
the equals sign is used as part of a comparison operator, then it is NOT
a delimeter between the param name and the expression.  rather it is
part of the expression.

This commit updates the code to account for this.