Make libxyce a monolithic library again
Many years ago, all subdirectories of Xyce were assembled into a
single library, libxyce.  Each subdirectory was built as a
"convenience library", but at the end of the build all the convenience
libraries were collected together into one libxyce that would be
installable, and that would be linked into a Xyce binary as well.

Almost immediately after inventing the plugin system
that allows Verilog-A models to be imported at run time, a
half-hearted attempt was made to extend the plugin system to apply to
optional model subsets (ADMS, non-free, neuron models, etc.).  The
idea was that there would be a SINGLE Xyce build of the main code, and
then all model libraries would be plugins.  That way we could
distribute non-free or OUO libraries together with the normal build
and automatically have an extensible Xyce.

The attempt was half-hearted because only the very first step was ever
taken: making all the optional models into separate libraries.  No
effort was ever taken to create plugins out of them, and in fact the
libraries so created had circular dependencies (libxyce depended on
register functions in the model libraries, and the model libraries all
depended on functions libxyce).

That developer is no longer with us.

What happened around Xyce 6.1 was that the formerly "noinst"
convenience libraries for SOME subdirectories were elevated to
installed libraries, and were no longer pulled into the monolithic
libxyce.

We left it this way because our own use cases never seemed to be
broken by this method, but it was always pointless given that the
reason it was done was never completed (or even started, really), and
wasn't really even an important thing to do.

Now we find that this unnecessary and poorly executed separation of
functions is impacting open source users who are trying to use Xyce
libraries in novel ways.  The circular dependencies are a problem, and
the fact that the libraries are not created by linking them completely
means that dynamic loaders can have trouble resolving symbols.

This commit undoes that ill-considered change and reverts to what we
always did before Xyce 6.1: all subdirectories now create convenience
libraries, and they are all pulled into libxyce.

This change means that linking to Xyce libraries is now simpler: you
only need the one, no matter which Xyce build you're using (Xyce,
XyceNF, XyceRad, etc.).

The two Jenkins jobs that build shared Xyce and install it, then
attempt to test linking to those libraries will also need changing so
they only link libxyce and not all the now-nonexistent libraries.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/204