Partial fix for bug 1145 SON
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1145
".GLOBAL_PARAM does not work in .IC statements"

This fix *completely*  fixes the issue for transient, but only
partially fixes the issue for HB.  It works fine with HB, but
only if the global parameter isn't swept as part of a .STEP.

The only substantive code change in this commit is in the
N_IO_InitialConditions.C file.  The other two files are only some
small formatting changes, that I noticed as I browsed them to
better understand what they did under .STEP.

In the N_IO_InitialConditions.C file,
InitialConditionsManager::setupInitialConditions function is what gets
changed in this commit, and the main thing that I had to change was a line in
that function that called the getImmutableValue function on an IC parameter.

That line only called getImmutableValue if the parameter returned
"false" for the function hasExpressionValue().  Presumably this was to
ensure that the value for .IC could never be changed, meaning that
global_params were not allowed there.  The thing is, I think that this
"getImmutableValue" function is over-used in Xyce.  In many cases, a
changing value could work just fine, and we're just assuming it won't
work without really double-checking if this is true.  I think that was
the case for this bug.

Anyway, I modified this line to instead call the new
"Param::getMutableValue" function.  As long as the expression can be
evaluated at this stage, this works fine.  This change allows Xyce
to get thru the parsing and setup of .IC, even when it depends on
a global parameter.

This commit fixes the issue for all analysis types, as long as the
global_parameter isn't swept by .STEP.  If that global_paramer is swept
by a .STEP (or .SAMPLING, etc) then some analysis types (like
transient) do the right thing, but some do not (like HB).

Using .STEP on the global_param used by the .IC statement works for
transient (and probably DC and AC).  HB still has some issues and
doens't quite work in that case.  For HB, it seems to produce the same
final HB answer at every step, even though the initial condition
transient does change correctly.  So, more work is needed there, but
hopefully not a lot.