More updates to the CMake build instructions
This reorders things to an extent, and adds a "TL;DR" section for those
who just want to type in some lines and get a result. If they get the
TPLs via a package manager, it should work. I also simplified the MacOS
instructions, getting rid of some unnecessary advice.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/1