Fix issue #206, for model name confusion
Issue #206.
"Parser gets confused with node names that match model names (at least
in one example)"

Also Issue #144, indirectly.
"Make model binning capability turned on by default"

Between Xyce 7.2 and Xyce 7.3, the logic for resolving model names in
the device block got broken.   It got broken in a subtle way that was
not caught by the regression test suite, but did get caught recently in
a customer circuit.

It got broken when I refactored Xyce to have modeling binning turned
on by default. (Issue #144)   As part of that work I
rewrote the DeviceBlock::extractModelName function.  But, I made a logic
mistake.

The extractModelName function checks if a "found" model is appropriate
for the device in question.  It checks a bunch of things, such as the
location on the instance line, as well as the type of model.   When I
refactored this function, I moved this check at the very end of the
function, but I needed to do it two times.   In other words, I needed to
keep the check where it was originally, and then do it again later
if binning is used.

Sincen the model binning is only executed  only if the initial
block has failed to find a valid model, the valid model check needs
to be done in both places.

Since it has to be done twice, I've created a new function to do this
checking.  This commit implements the new function, and calls it
from both places that it needs to be called.  That should mostly
fix the issue.